In model clear cache.

+++ app/Models/Permission.php
@@ -24,14 +24,6 @@ class Permission extends Model
     {
         parent::boot();
 
-
-        // При создании элемента
-        static::creating(function (self $model) {
-
-            // Удалить кэш
-            cache()->flush();
-        });
-
         // При сохранении элемента
         static::saving(function (self $model) {
 
@@ -42,8 +34,12 @@ class Permission extends Model
         // При удалении элемента
         static::deleting(function (self $model) {
 
-            // Удаление связанного элемента
-            $model->permissions()->delete();
+            // При мягком удалении не сработает
+            if ($model->isForceDeleting()) {
+
+                // Удаление связанного элемента
+                $model->permissions()->delete();
+            }
 
             // Удалить кэш
             cache()->flush();
