+++ app/Models/Role.php
@@ -55,14 +55,6 @@ class Role extends Model implements ModelPermissionInterface
     {
         parent::boot();
 
-
-        // При создании элемента
-        static::creating(function (self $model) {
-
-            // Удалить кэш
-            cache()->flush();
-        });
-
         // При сохранении элемента
         static::saving(function (self $model) {
 
@@ -73,8 +65,12 @@ class Role extends Model implements ModelPermissionInterface
         // При удалении элемента
         static::deleting(function (self $model) {
 
-            // Удаление связанного элемента
-            $model->permissions()->delete();
+            // При мягком удалении не сработает
+            if ($model->isForceDeleting()) {
+
+                // Удаление связанного элемента
+                $model->permissions()->delete();
+            }
 
             // Удалить кэш
             cache()->flush();
